@@ -15,44 +15,48 @@
   ],
   "dependencies": {
     "asn1.js": "^4.9.0",
-    "tweetnacl": "^0.14.1"
+    "tweetnacl": "^1.0.0"
   },
   "devDependencies": {
-    "babel-loader": "^6.2.7",
+    "ajv": "^5.3.0",
+    "babel-cli": "^6.26.0",
+    "babel-loader": "^7.1.2",
     "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-transform-runtime": "^6.7.5",
-    "babel-polyfill": "^6.16.0",
-    "babel-preset-env": "^1.6.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-es2015": "^6.9.0",
-    "babel-runtime": "^6.9.0",
-    "babelify": "^7.2.0",
+    "babel-runtime": "^6.26.0",
+    "babelify": "^8.0.0",
     "brfs": "^1.4.3",
-    "chai": "^3.2.0",
-    "codecov": "^1.0.1",
-    "cz-conventional-changelog": "^1.1.5",
-    "eslint": "^3.1.1",
-    "eslint-config-standard": "^6.0.0",
-    "eslint-plugin-promise": "^3.0.0",
-    "eslint-plugin-standard": "^2.0.0",
-    "husky": "^0.12.0",
+    "chai": "^4.1.2",
+    "codecov": "^3.0.0",
+    "cz-conventional-changelog": "^2.1.0",
+    "eslint": "^4.11.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^5.2.1",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-standard": "^3.0.1",
+    "husky": "^0.14.3",
     "interledger-jsdoc-template": "^2.0.0",
-    "jsdoc": "^3.4.0",
-    "json-loader": "^0.5.4",
-    "karma": "^1.1.0",
-    "karma-mocha": "^1.1.0",
-    "karma-phantomjs-launcher": "^1.0.0",
+    "jsdoc": "^3.5.5",
+    "json-loader": "^0.5.7",
+    "karma": "^1.7.1",
+    "karma-mocha": "^1.3.0",
+    "karma-phantomjs-launcher": "^1.0.4",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^1.8.0",
+    "karma-webpack": "^2.0.6",
     "md-toc-filter": "^0.9.0",
-    "mocha": "^3.0.0",
+    "mocha": "^4.0.1",
     "mustache": "^2.3.0",
-    "nyc": "^10.0.0",
-    "phantomjs-prebuilt": "^2.1.5",
+    "nyc": "^11.3.0",
+    "phantomjs-prebuilt": "^2.1.16",
     "readme-to-test": "justmoon/readme-to-test#feature/st-improvements",
-    "release-it": "^3.0.2",
-    "sinon": "^1.17.4",
-    "watchify": "^3.7.0",
-    "webpack": "^2.1.0-beta.25"
+    "sinon": "^4.1.2",
+    "watchify": "^3.9.0",
+    "webpack": "^3.8.1"
+    "release-it": "^5.0.0"
   },
   "scripts": {
     "lint": "eslint .",
